removed applu.i and maxnorm.i from applu file list
these files do not show up in the exec.stats when running the applu
benchmark. i suspect it is because these functions are not called.
They are removed from the file list so that no searching is done
(also it was causing an exception because the results were not applied
to the correct search state )

darcs-hash:20070724134648-22c1e-535747cb9f6073d1efdbb6f3012399b5e4ab74db